Rearrange folders and files in components for easy redability
